+++ starsky/starsky/clientapp/src/components/modal-export.spec.tsx
@@ -1,8 +1,10 @@
+import { act } from '@testing-library/react';
 import { mount, shallow } from 'enzyme';
 import React from 'react';
-import { act } from 'react-dom/test-utils';
 import * as useFetch from '../hooks/use-fetch';
+import * as useInterval from '../hooks/use-interval';
 import { IConnectionDefault } from '../interfaces/IConnectionDefault';
+import * as FetchPost from '../shared/fetch-post';
 import * as Modal from './modal';
 import ModalExport from './modal-export';
 
@@ -41,6 +43,44 @@ describe("ModalExport", () => {
     });
   });
 
+  it("Multiple Files -> click download ", () => {
+    // use ==> import * as useFetch from '../hooks/use-fetch';
+    const mockGetIConnectionDefault = {
+      statusCode: 200, data: null
+    } as IConnectionDefault;
+
+    jest.spyOn(useFetch, 'default')
+      .mockImplementationOnce(() => mockGetIConnectionDefault)
+      .mockImplementationOnce(() => mockGetIConnectionDefault)
+
+    jest.spyOn(useInterval, 'default').mockImplementationOnce(() => { })
+      .mockImplementationOnce(() => { })
+
+    // use ==> import * as FetchGet from '../shared/fetch-get';
+    const mockFetchGetIConnectionDefault: Promise<IConnectionDefault> = Promise.resolve({
+      statusCode: 200, data: null
+    } as IConnectionDefault);
+
+    var fetchPostSpy = jest.spyOn(FetchPost, 'default')
+      .mockImplementationOnce(() => mockFetchGetIConnectionDefault)
+
+    var modal = mount(<ModalExport select={["/file0", "/file1.jpg"]} isOpen={true} handleExit={() => { }}></ModalExport>)
+
+    var item = modal.find('[data-test="thumbnail"]');
+
+    act(() => {
+      item.simulate('click');
+    });
+
+    expect(fetchPostSpy).toBeCalled();
+
+    // and clean afterwards
+    jest.spyOn(window, 'scrollTo').mockImplementationOnce(() => { });
+    act(() => {
+      modal.unmount();
+    });
+  });
+
   it("file type not supported", () => {
     // use ==> import * as useFetch from '../hooks/use-fetch';
     const mockGetIConnectionDefault = {
@@ -70,14 +110,26 @@ describe("ModalExport", () => {
       return <>{props.children}</>
     });
 
+    jest.spyOn(useInterval, 'default').mockImplementationOnce(() => { });
+
+    // use ==> import * as useFetch from '../hooks/use-fetch';
+    const mockGetIConnectionDefault = {
+      statusCode: 415, data: null
+    } as IConnectionDefault;
+    jest.spyOn(useFetch, 'default')
+      .mockImplementationOnce(() => mockGetIConnectionDefault)
+
     var handleExitSpy = jest.fn();
 
-    var component = mount(<ModalExport select={["/"]} isOpen={true} handleExit={handleExitSpy} />);
+    var modal = mount(<ModalExport select={["/"]} isOpen={true} handleExit={handleExitSpy} />);
 
     expect(handleExitSpy).toBeCalled();
 
     // and clean afterwards
-    component.unmount();
+    jest.spyOn(window, 'scrollTo').mockImplementationOnce(() => { });
+    act(() => {
+      modal.unmount();
+    });
   });